kernel/fla: export per-chunk GDN state h in fp32 (P3-29 cold/resume divergence)
The hybrid-radix track checkpoint snapshots h into the fp32 state pool, but
h was bf16 while the cold path carries the inter-chunk state in fp32
registers -- every snapshot resume restarted from a rounded state, the
delta=1e-2 token-0 logprob divergence between cold prefill and radix+GDN
resume found in P2 smoke. h is now fp32; pass 2 (chunk_fwd_o) casts it back
to bf16 at the dot, so prefill outputs are bit-identical to the old kernel
and only the snapshot gains the exact register state. Resumed continuations
now reproduce a cold full prefill bit for bit (outputs and final state);
pinned by tests/kernels/test_gdn_chunk_snapshot.py including a sensitivity
probe that a bf16-rounded snapshot does shift the continuation.

+++ python/freetoken/kernel/fla/chunk_delta_h.py
@@ -321,7 +321,10 @@ def chunk_gated_delta_rule_fwd_h(
         )
     assert K <= 256, "current kernel does not support head dimension larger than 256."
 
-    h = k.new_empty(B, NT, H, V, K)
+    # fp32, not k.dtype: h feeds the hybrid-radix track snapshots, and the cold
+    # path carries the inter-chunk state in fp32 registers -- a bf16 h restarts a
+    # resumed continuation from a rounded state (P3-29 cold-vs-resume divergence).
+    h = k.new_empty(B, NT, H, V, K, dtype=torch.float32)
 
     v_new = torch.empty_like(u) if save_new_value else None
